Move DiagramHelper.ViewTypes to separate file & delete RepesentationKind
