+++ plugins/org.eclipse.papyrus.uml.diagram.sequence.runtime/src/org/eclipse/papyrus/uml/diagram/sequence/runtime/internal/preferences/LightweightSequenceDiagramPreferencePage.java
@@ -14,7 +14,7 @@ package org.eclipse.papyrus.uml.diagram.sequence.runtime.internal.preferences;
 
 import org.eclipse.papyrus.infra.gmfdiag.preferences.pages.DiagramPreferencePage;
 import org.eclipse.papyrus.uml.diagram.sequence.runtime.internal.Activator;
-import org.eclipse.papyrus.uml.interaction.model.spi.DiagramHelper.ViewTypes;
+import org.eclipse.papyrus.uml.interaction.model.spi.ViewTypes;
 
 /**
  * This is the {@code LightweightSequenceDiagramPreferencePage} type. Enjoy.
