+++ plugins/org.eclipse.papyrus.uml.interaction.model.view/src/org/eclipse/papyrus/uml/interaction/model/internal/view/LogicalModelView.java
@@ -15,7 +15,7 @@ package org.eclipse.papyrus.uml.interaction.model.internal.view;
 
 import org.eclipse.gmf.runtime.notation.Diagram;
 import org.eclipse.papyrus.uml.diagram.common.part.UmlGmfDiagramEditor;
-import org.eclipse.papyrus.uml.interaction.model.spi.DiagramHelper.ViewTypes;
+import org.eclipse.papyrus.uml.interaction.model.spi.ViewTypes;
 import org.eclipse.ui.part.IPage;
 import org.eclipse.ui.part.IPageBookViewPage;
